+++ cisst/cisstStereoVision/code/svlSampleCameraGeometry.cpp
@@ -157,7 +157,7 @@ svlSampleCameraGeometry::svlSampleCameraGeometry(const std::string & calibration
 
 unsigned int svlSampleCameraGeometry::GetCameraCount() const
 {
-    return IntrinsicVector.size();
+    return static_cast<unsigned int>(IntrinsicVector.size());
 }
 
 int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
@@ -184,7 +184,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
     try {
         fin.open(filepath.c_str());
         memset(chstr, 0, 2048);
-        pos = fin.read(chstr, 2047).gcount();
+        pos = static_cast<int>(fin.read(chstr, 2047).gcount());
     } catch (std::ifstream::failure e) {
     }
     fin.close();
@@ -199,7 +199,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         // LEFT CAMERA INTRINSIC
 
         // left focal length
-        pos = str.find("fc_left = [ ");
+        pos = static_cast<int>(str.find("fc_left = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 12);
             strstrm >> fc[0] >> fc[1];
@@ -207,7 +207,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         else success = SVL_FAIL;
 
         // left principal point
-        pos = str.find("cc_left = [ ");
+        pos = static_cast<int>(str.find("cc_left = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 12);
             strstrm >> cc[0] >> cc[1];
@@ -215,7 +215,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         else success = SVL_FAIL;
 
         // left skew
-        pos = str.find("alpha_c_left = [ ");
+        pos = static_cast<int>(str.find("alpha_c_left = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 17);
             strstrm >> a;
@@ -223,7 +223,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         else success = SVL_FAIL;
 
         // left radial distortion
-        pos = str.find("kc_left = [ ");
+        pos = static_cast<int>(str.find("kc_left = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 12);
             strstrm >> kc[0] >> kc[1] >> kc[2] >> kc[3] >> kc[4] >> kc[5] >> kc[6];
@@ -236,7 +236,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         // RIGHT CAMERA INTRINSIC
 
         // right focal length
-        pos = str.find("fc_right = [ ");
+        pos = static_cast<int>(str.find("fc_right = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 13);
             strstrm >> fc[0] >> fc[1];
@@ -244,7 +244,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         else success = SVL_FAIL;
 
         // right principal point
-        pos = str.find("cc_right = [ ");
+        pos = static_cast<int>(str.find("cc_right = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 13);
             strstrm >> cc[0] >> cc[1];
@@ -252,7 +252,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         else success = SVL_FAIL;
 
         // right skew
-        pos = str.find("alpha_c_right = [ ");
+        pos = static_cast<int>(str.find("alpha_c_right = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 18);
             strstrm >> a;
@@ -260,7 +260,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         else success = SVL_FAIL;
 
         // right radial distortion
-        pos = str.find("kc_right = [ ");
+        pos = static_cast<int>(str.find("kc_right = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 13);
             strstrm >> kc[0] >> kc[1] >> kc[2] >> kc[3] >> kc[4] >> kc[5] >> kc[6];
@@ -276,7 +276,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         SetExtrinsics(0.0, 0.0, 0.0, 0.0, 0.0, 0.0, SVL_LEFT);
 
         // rotation between cameras
-        pos = str.find(" om = [ ");
+        pos = static_cast<int>(str.find(" om = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 8);
             strstrm >> om[0] >> om[1] >> om[2];
@@ -284,7 +284,7 @@ int svlSampleCameraGeometry::LoadCalibration(const std::string & filepath)
         else success = SVL_FAIL;
 
         // translation between cameras
-        pos = str.find(" T = [ ");
+        pos = static_cast<int>(str.find(" T = [ "));
         if (pos >= 0) {
             strstrm.seekg(pos + 7);
             strstrm >> T[0] >> T[1] >> T[2];
@@ -539,7 +539,7 @@ int svlSampleCameraGeometry::IsCameraPairRectified(const unsigned int cam_id1, c
 int svlSampleCameraGeometry::SetWorldToCenter()
 {
     unsigned int i;
-    const unsigned int camcount = ExtrinsicVector.size();
+    const unsigned int camcount = static_cast<unsigned int>(ExtrinsicVector.size());
     if (camcount == 0) return SVL_FAIL;
     if (camcount == 1) return SVL_OK;
 
@@ -577,7 +577,7 @@ int svlSampleCameraGeometry::SetWorldToCenter()
 
 int svlSampleCameraGeometry::RotateWorldAboutY(double degrees)
 {
-    const unsigned int camcount = ExtrinsicVector.size();
+    const unsigned int camcount = static_cast<unsigned int>(ExtrinsicVector.size());
     if (camcount == 0) return SVL_FAIL;
 
     // Create axis-angle rotation
@@ -603,7 +603,7 @@ int svlSampleCameraGeometry::RotateWorldAboutY(double degrees)
 
 int svlSampleCameraGeometry::RotateWorldAboutZ(double degrees)
 {
-    const unsigned int camcount = ExtrinsicVector.size();
+    const unsigned int camcount = static_cast<unsigned int>(ExtrinsicVector.size());
     if (camcount == 0) return SVL_FAIL;
 
     // Create axis-angle rotation
